dbs_backend_flask/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
pymysql
flask-cors